refactor(desktop): re-fit voice store to createStoreContext
Voice now matches the chat store's pattern: a Store provided via
createStoreContext (VoiceProvider at the app root), actions on the store via
the factory, consumers on useSelector + store.actions. Server state un-mirrored
per the Query/Store split — voices + health move out of the store into
useVoiceHealth() (a thin wrapper over the voices query). voice.ts -> voice.tsx
(now renders a provider). Both desktop features are on one pattern.
