+++ apps/desktop/src/mainview/app/index.tsx
@@ -2,20 +2,15 @@ import { QueryClientProvider } from "@tanstack/react-query";
 import { RouterProvider } from "@tanstack/react-router";
 
 import { queryClient } from "~/lib/query-client";
-import { useVoiceRuntime } from "~/stores/voice";
+import { VoiceProvider } from "~/stores/voice";
 import { router } from "./router";
-
-function AppShell() {
-  // Mount the voice runtime once, inside the Query provider, so it can drive
-  // the voices health query and hydrate the voice store.
-  useVoiceRuntime();
-  return <RouterProvider router={router} />;
-}
 
 function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <AppShell />
+      <VoiceProvider>
+        <RouterProvider router={router} />
+      </VoiceProvider>
     </QueryClientProvider>
   );
 }
