+++ apps/desktop/src/mainview/app/screens/chat/components/top-bar.tsx
@@ -1,10 +1,9 @@
-import { useSelector } from "@tanstack/react-store";
 import { Settings } from "lucide-react";
 import { match } from "ts-pattern";
 
 import { ModelPicker } from "~/app/components/model-picker";
 import { DRAG, NO_DRAG } from "~/lib/drag-region";
-import { voiceStore } from "~/stores/voice";
+import { useVoiceHealth } from "~/stores/voice";
 import { Button } from "~/ui/button";
 import { SidebarTrigger } from "~/ui/sidebar";
 import { Tooltip, TooltipContent, TooltipTrigger } from "~/ui/tooltip";
@@ -16,7 +15,7 @@ interface ChatTopBarProps {
 }
 
 export function ChatTopBar({ model, onModelChange }: ChatTopBarProps) {
-  const health = useSelector(voiceStore, (s) => s.health);
+  const { health } = useVoiceHealth();
   const isHealthy = health.kind === "ok";
 
   return (
